+++ src/app/(main)/collections/[handle]/loading.tsx
