+++ src/app/(main)/page.tsx
@@ -1,4 +1,3 @@
-import PageLayout from "./page-layout"
 import FeaturedProducts from "@modules/home/components/featured-products"
 import Hero from "@modules/home/components/hero"
 import { Metadata } from "next"
@@ -11,10 +10,10 @@ export const metadata: Metadata = {
 
 const Home = () => {
   return (
-    <PageLayout>
+    <>
       <Hero />
       <FeaturedProducts />
-    </PageLayout>
+    </>
   )
 }
