+++ src/app/(main)/products/[handle]/page.tsx
