--- src/app/account/profile/layout.tsx
@@ -1,9 +1,0 @@
-import AccountLayout from "@modules/account/templates/account-layout"
-
-export default function ProfilePageLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <AccountLayout>{children}</AccountLayout>
-}
