--- src/app/order/layout.tsx
@@ -1,9 +1,0 @@
-import PageLayout from "app/page-layout"
-
-export default function CollectionLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <PageLayout>{children}</PageLayout>
-}
