--- src/app/products/[handle]/layout.tsx
@@ -1,9 +1,0 @@
-import PageLayout from "app/page-layout"
-
-export default function ProductLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <PageLayout>{children}</PageLayout>
-}
