--- src/app/store/layout.tsx
@@ -1,9 +1,0 @@
-import PageLayout from "app/page-layout"
-
-export default function StoreLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <PageLayout>{children}</PageLayout>
-}
